+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Next, enable the PSPDFKit.xcconfig project configuration. Here's a screenshot: h
 *  PSPDFViewController now has a margin and a padding property to add custom margin/padding on the pdf view.
 *  PSPDFViewController now has a HUDViewMode property to fine-tune the HUD.
 *  PSPDFViewController now has a renderingMode property; fine-tune cache/thumbnail view usage. PSPDFPageRenderingFullPageBlocking is *great *for pageCurl mode.
+*  New delegate: - (void)pdfViewController:(PSPDFViewController *)pdfController didEndPageDragging:(UIScrollView *)scrollView willDecelerate:(BOOL)decelerate withVelocity:(CGPoint)velocity targetContentOffset:(inout CGPoint *)targetContentOffset -> completes dragging delegates; zooming delegates were already available in PSPDFKit v1.
 *  Add overrideClassNames to PSPDFDocument.
 *  Lots of other minor and major changes.