+++ CHANGELOG.md
@@ -38,7 +38,10 @@ Next, enable the PSPDFKit.xcconfig project configuration. Here's a screenshot: h
 *  Table of Contents now shows titles in multiple lines if too long.
 *  The PSPDFKit image annotation now properly displays and animates animated GIFs.
 *  A long-press on a PDF link annotation now shows the URL/Document/Page target in a popover.
+*  The PSPDFScrobbleBar now uses the small height style on iPhone/Landscape.
 *  Fixes a problem in the annotation parser where some named page links failed to resolve properly.
+*  Fixes a problem where the Cancel button of the additionalBarButtonItem menu wasn't fully touch-able on iPhone.
+*  Fixes a problem with PSPDFBarButtonItem image updating.
 *  Lots of other minor and major changes.
 
 __v1.10.4 - x/Jul/2012__